Move examples out of the main repo to twbs/bootstrap-examples
While examples are phenomenally useful and awesome for folks getting
started with Bootstrap, they don't really shine as well being buried in
the main docs. Moreover, having anything that's not core HTML, CSS, JS,
or documentation in the main repo seems to just get in the way.
Therefore, examples gets its own repo.

Specifically, we're doing the following:

* Deletes all example screenshots
* Deletes example layout
* Deletes example files
* Updates Getting Started docs to include only the template and a link
to the new examples project

All the examples and their history will be preserved in the switch.