Fix holder.js links in examples
Replaces links to holder.js with links to the new docs.min.js.
The uncompressed holder.js file is now in docs/assets/js/vendor,
which is excluded from Jekyll builds. In turn, Holder is now
included in docs.min.js, which is smaller than holder.js,
so this should make everybody happy. :)